requirements.txt
flask
openai-whisper
torch
transformers
sentencepiece
ffmpeg-python